delete old qio::vcf package
